respawn point push

+++ programs/orbit-physical-market/src/structs/physical_product.rs
@@ -4,5 +4,5 @@ use anchor_lang::prelude::*;
 #[account]
 pub struct PhysicalProduct{
     pub metadata: OrbitProduct, // 38 as of beta
-    pub quantity: i8, // quantity per purchase // 1
+    pub quantity: u32, // quantity per purchase // 4
 }
